PlayRegisterを軽くする
